19: Added new API to create datasource Object
Using this API, user can obtain a datasource object by passing a datasourcedefinition
